jobs/nro-extractor/requirements/prod.txt
Flask
Flask-SQLAlchemy
Flask-Restplus
Flask-Marshmallow

git+https://github.com/bcgov/namex.git#egg=namex&subdirectory=api

psycopg2-binary
python-dotenv
marshmallow-sqlalchemy
gunicorn
gunicorn[eventlet]
greenlet
eventlet
